use separator
